Added routes-directory and fixed quotes routing
