+++ Server/server.js
@@ -1,7 +1,9 @@
 //Imports
 const path = require('path');
-const fs = require('fs');
 const express = require('express');
+
+//Route to motivational quotes
+const quoteRouter = require('./routes/quoteRoutes');
 
 //Create instance of express application
 const app = express();
@@ -15,22 +17,7 @@ app.get('/topnavbar.html', (req, res) => {
     res.sendFile('C:/work/image-rec-proj/Client/topnavbar.html');
 });
 
-// Read the "random-quotes.txt" file and store quotes in memory during server startup
-let randomQuotes = [];
-fs.readFile('./Server/random-quotes.txt', 'utf8', (error, data) => {
-    if (error) {
-        console.log('Error reading file:', error);
-        return;
-    }
-    const randomQuotesDocument = data.toString();
-    randomQuotes = randomQuotesDocument.split('\n').map(quote => quote.trim());
-});
-
-// API endpoint for a random quote
-app.get('/api/random-quote', (req, res) => {
-    const randomIndex = Math.floor(Math.random() * randomQuotes.length);
-    const randomQuote = randomQuotes[randomIndex];
-    res.json({ text: randomQuote });
-});
+//use the quotesRoute
+app.use('/', quoteRouter);
 
 module.exports = app; // Export the 'app' variable for unit-tests
